Update initialState of Diagram/index.js to use props
Now the data can be set while creating the the diagram. The default
values can still be used if no other data is provided. These default
values should however be removed as fast correct formatted data is
available as there would no longer be any benefit from having sample
data.

Setting the state from props could be considered anti-pattern if it
would not be because of the fact that state is used to render the
diagram.